refactor: 関数名の修正
タグと手入力の情報を一括で管理するように変更したため
