+++ src/components/index.ts
@@ -19,7 +19,7 @@ export { default as MarkdownPreview } from "./common_part/MarkdownPreview";
 export { default as DisplayProduct } from "./common_part/DisplayProduct";
 export { default as DisplayProducts } from "./common_part/DisplayProducts";
 export { default as Like } from "./common_part/Like";
-export { default as DisplayTagList } from "./Home/DisplayTagList";
+export { default as SearchCondition } from "./Home/SearchCondition";
 export { default as postImage } from "./common_part/postImage";
 export type { default as DisplayProductProps } from "./common_part/DisplayProductProps";
 export { default as DisplayUserInfo } from "./user/DisplayUserInfo";
